backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
mysql-connector-python==8.2.0
aiomysql==0.2.0

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Gemini AI
google-generativeai==0.3.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Async
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1